strategy_agent/requirements.txt
flask
gunicorn
google-cloud-pubsub
google-cloud-firestore
